backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
databricks-sdk>=0.18.0
databricks-sql-connector>=3.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
cachetools>=5.3.0
python-multipart>=0.0.6
